fix: 리듀서 모듈 분리
